Revert "Improvement/remove items by quantity"
