+++ Tomi.Application/ApiResponse/Response.cs
@@ -5,17 +5,17 @@
         public Response()
         {
         }
-        public Response(T data, bool succeeded, string message = null)
+        public Response(T data, string message = null)
         {
-            Succeeded = succeeded;
+            Succeeded = true;
             Message = message;
             Data = data;
         }
-		public Response(bool succeeded, string message = null)
-		{
-			Succeeded = succeeded;
-			Message = message;
-		}
+        public Response(string message)
+        {
+            Succeeded = false;
+            Message = message;
+        }
         public bool Succeeded { get; set; }
         public string Message { get; set; }
         public List<string> Errors { get; set; }
