--- Tomi.Application/Features/Products/Queries/GetAll/GetAllProductsHandler.cs
@@ -1,32 +1,0 @@
-﻿using AutoMapper;
-using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Tomi.Application.Models;
-using Tomi.Domain.Entities;
-using Tomi.Domain.IRepositories;
-
-namespace Tomi.Application.Features.Products.Queries.GetAll
-{
-    public class GetAllProductsHandler : IRequestHandler<GetAllProductsQuery, IEnumerable<ProductModel>>
-    {
-        private readonly IProductRepository _productRepository;
-        private readonly IMapper _mapper;
-
-        public GetAllProductsHandler(IProductRepository productRepository, IMapper mapper)
-        {
-            _productRepository = productRepository;
-            _mapper = mapper;
-        }
-
-        public async Task<IEnumerable<ProductModel>> Handle(GetAllProductsQuery request, CancellationToken cancellationToken)
-        {
-            var products = await _productRepository.GetAllAsync();
-            var productViewModel = _mapper.Map<IEnumerable<ProductModel>>(products);
-            return productViewModel;
-        }
-    }
-}
