+++ Tomi.Application/Features/Products/Queries/GetAllProductsQuery.cs
@@ -1,10 +1,10 @@
 ﻿using MediatR;
 using Tomi.Application.Models;
 
-namespace Tomi.Application.Features.Products.Queries.GetAll
+namespace Tomi.Application.Features.Products.Queries
 {
     public class GetAllProductsQuery : IRequest<IEnumerable<ProductModel>>
-    {
-        //open for extension
-    }
+	{
+		//open for extension
+	}
 }
