+++ Tomi.Application/Features/Products/Queries/GetProductByIdQuery.cs
@@ -6,9 +6,10 @@ using System.Text;
 using System.Threading.Tasks;
 using Tomi.Application.Models;
 
-namespace Tomi.Application.Features.Coupons.Queries.GetAll
+namespace Tomi.Application.Features.Products.Queries
 {
-    public class GetAllCouponsQuery : IRequest<IEnumerable<CouponModel>>
-    {
-    }
+    public class GetProductByIdQuery : IRequest<ProductModel>
+	{
+		public string Id { get; set; }
+	}
 }
