--- Tomi.Application/Features/ShoppingCarts/Commands/ClearShoppingCart/ClearShoppingCartCommand.cs
@@ -1,10 +1,0 @@
-﻿using MediatR;
-using Tomi.Application.ApiResponse;
-
-namespace Tomi.Application.Features.ShoppingCarts.Commands.RemoveAllItem
-{
-    public class ClearShoppingCartCommand : IRequest<Response<string>>
-    {
-        public string UserId { get; set; }
-    }
-}
