+++ Tomi.Application/Services/Handlers/Coupons/ApplyCouponHandler.cs
@@ -1,11 +1,12 @@
 ﻿using AutoMapper;
 using MediatR;
 using Tomi.Domain.Enums.CouponEnums;
+using Tomi.Application.Features.Coupons.Commands;
 using Tomi.Application.Models;
 using Tomi.Domain.Entities;
 using Tomi.Domain.IRepositories;
 
-namespace Tomi.Application.Features.Coupons.Commands.Apply
+namespace Tomi.Application.Services.Handlers.Coupons
 {
     public class ApplyCouponHandler : IRequestHandler<ApplyCouponCommand, ApplyCouponModel>
     {
