+++ Tomi.Application/Services/Handlers/Products/GetProductByIdHandler.cs
@@ -1,0 +1,33 @@
+﻿using AutoMapper;
+using MediatR;
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Text;
+using System.Threading.Tasks;
+using Tomi.Application.Features.Products.Queries;
+using Tomi.Application.Models;
+using Tomi.Domain.Entities;
+using Tomi.Domain.IRepositories;
+
+namespace Tomi.Application.Services.Handlers.Products
+{
+    public class GetProductByIdHandler : IRequestHandler<GetProductByIdQuery, ProductModel>
+	{
+		private readonly IProductRepository _productRepository;
+		private readonly IMapper _mapper;
+		public GetProductByIdHandler(IProductRepository productRepository, IMapper mapper) { 
+		    _productRepository = productRepository;
+			_mapper = mapper;
+		}
+		public async Task<ProductModel> Handle(GetProductByIdQuery request, CancellationToken cancellationToken)
+		{
+			var product = await _productRepository.GetByIdAsync(request.Id);
+			var productViewModel = _mapper.Map<ProductModel>(product);
+			return productViewModel;
+
+		}
+	}
+}
