+++ Tomi.Application/Services/Handlers/ShoppingCarts/RemoveItemFromShoppingCart.cs
@@ -1,27 +1,26 @@
 ﻿using AutoMapper;
 using MediatR;
-using Tomi.Application.ApiResponse;
-using Tomi.Application.Auth.Models;
+using Tomi.Application.Features.ShoppingCarts.Commands;
 using Tomi.Application.Models;
 using Tomi.Domain.Entities;
 using Tomi.Domain.IRepositories;
 
-namespace Tomi.Application.Features.ShoppingCarts.Commands.RemoveItem
+namespace Tomi.Application.Services.Handlers.ShoppingCarts
 {
-    public class RemoveItemFromShoppingCartHandler : IRequestHandler<RemoveItemCommand, Response<string>>
+    public class RemoveItemFromShoppingCart : IRequestHandler<RemoveItemCommand, ShoppingCartItemModel>
     {
         private readonly IShoppingCartRepository _shoppingCartRepository;
         private readonly IProductRepository _productRepository;
         private readonly IMapper _mapper;
 
-        public RemoveItemFromShoppingCartHandler(IShoppingCartRepository shoppingCartRepository, IProductRepository productRepository, IMapper mapper)
+        public RemoveItemFromShoppingCart(IShoppingCartRepository shoppingCartRepository, IProductRepository productRepository, IMapper mapper)
         {
             _shoppingCartRepository = shoppingCartRepository;
             _productRepository = productRepository;
             _mapper = mapper;
         }
 
-        public async Task<Response<string>> Handle(RemoveItemCommand request, CancellationToken cancellationToken)
+        public async Task<ShoppingCartItemModel> Handle(RemoveItemCommand request, CancellationToken cancellationToken)
         {
             var shoppingCart = await _shoppingCartRepository.GetByUserIdAsync(request.UserId);
             var product = await _productRepository.GetByIdAsync(request.ProductId);
@@ -53,7 +52,19 @@ namespace Tomi.Application.Features.ShoppingCarts.Commands.RemoveItem
             else
             {
                 await _shoppingCartRepository.UpdateAsync(shoppingCart.Id, shoppingCart);
-                return new Response<string>(true, message: "Item removed"); ;
+
+                var totalPrice = product.Price * existingCartItem.Count;
+                var totalCount = existingCartItem.Count;
+
+                var response = new ShoppingCartItemModel
+                {
+                    UserId = request.UserId,
+                    ProductId = product.Id,
+					ProductTotalPrice = totalPrice,
+                    ProductCount = totalCount
+                };
+
+                return response;
             }
         }
