+++ Tomi.Application/Tomi.Application.csproj
@@ -19,6 +19,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Tomi.Domain\Tomi.Domain.csproj" />
+    <ProjectReference Include="..\Tomi.Infrastructure\Tomi.Infrastructure.csproj" />
   </ItemGroup>
 
 </Project>
