+++ Tomi.Infrastructure/Repositories/BaseRepository.cs
@@ -1,11 +1,9 @@
-﻿
-using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Options;
 using MongoDB.Driver;
 using Tomi.Domain.Entities;
 using Tomi.Domain.IRepositories;
-using Tomi.Domain.Settings;
 using Tomi.Infrastructure.Contexts;
-
+using Tomi.Infrastructure.Settings;
 
 namespace Tomi.Infrastructure.Repositories
 {
