+++ Tomi.Infrastructure/Repositories/ProductRepository.cs
@@ -1,9 +1,8 @@
 ﻿using Microsoft.Extensions.Options;
 using Tomi.Domain.Entities;
 using Tomi.Domain.IRepositories;
-using Tomi.Domain.Settings;
 using Tomi.Infrastructure.Contexts;
-
+using Tomi.Infrastructure.Settings;
 
 namespace Tomi.Infrastructure.Repositories
 {
