+++ Tomi.Infrastructure/Services/TokenService.cs
@@ -1,16 +1,13 @@
 ﻿using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using System;
-using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
 using System.Security.Claims;
 using System.Text;
-using System.Threading.Tasks;
 using Tomi.Domain.Entities;
-using Tomi.Domain.Settings;
+using Tomi.Infrastructure.Settings;
 
-namespace Tomi.Application.Services
+namespace Tomi.Infrastructure.Services
 {
 	public class TokenService
 	{
