+++ Tomi.WebAPI/Controllers/ShoppingCartController.cs
@@ -2,12 +2,7 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Tomi.Application.Features.ShoppingCarts.Commands;
-using Tomi.Application.Features.ShoppingCarts.Commands.AddItem;
-using Tomi.Application.Features.ShoppingCarts.Commands.RemoveAllItem;
-using Tomi.Application.Features.ShoppingCarts.Commands.RemoveItem;
-using Tomi.Application.Features.ShoppingCarts.Commands.RemoveItemByQuantity;
 using Tomi.Application.Features.ShoppingCarts.Queries;
-using Tomi.Application.Features.ShoppingCarts.Queries.GetAll;
 using Tomi.Domain.Entities;
 
 namespace Tomi.WebAPI.Controllers
@@ -43,25 +38,20 @@ namespace Tomi.WebAPI.Controllers
 			return Ok(await Mediator.Send(command));
 		}
 
-
-		[HttpDelete("remove-item-by-quantity")]
-		public async Task<IActionResult> RemoveItemByQuantityFromShoppingCart(string productId, int quantity)
-		{
-			var command = new RemoveItemByQuantityCommand();
-			command.UserId = UserId;
-			command.ProductId = productId;
-			command.Quantity = quantity;
-			return Ok(await Mediator.Send(command));
-		}
-
-
-		[HttpDelete("clear-shopping-cart")]
-        public async Task<IActionResult> ClearShoppingCart()
+        [HttpDelete("clear-shopping-cart")]
+        public async Task<IActionResult> RemoveAllItemsFromShoppingCart()
         {
-			var command = new ClearShoppingCartCommand();
+			var command = new RemoveAllItemsCommand();
 			command.UserId = UserId;
 			return Ok(await Mediator.Send(command));
         }
-
+        [HttpDelete("remove-item-from-shopping-cart-compeletely")]
+        public async Task<IActionResult> RemoveItemFromShoppingCartCompeletely(string productId)
+        {
+            var command = new RemoveItemCompeletelyCommand();
+            command.UserId = UserId;
+            command.ProductId = productId;
+            return Ok(await Mediator.Send(command));
+        }
     }
 	}
